Move revocation interval from proofs to credentials
