+++ src/lib/modules/credentials/models/RevocationInterval.ts
