+++ src/lib/modules/credentials/models/index.ts
@@ -1,2 +1,3 @@
 export * from './Credential';
 export * from './CredentialInfo';
+export * from './RevocationInterval';
